+++ components/mail/Base.php
@@ -4,7 +4,7 @@ namespace app\components\mail;
 
 use app\components\HTMLTools;
 use app\models\settings\AntragsgruenApp;
-use app\models\db\{Consultation, EMailBlacklist, EMailLog};
+use app\models\db\{Consultation, EMailBlocklist, EMailLog};
 use app\models\exceptions\ServerConfiguration;
 use yii\helpers\Html;
 
@@ -131,8 +131,8 @@ abstract class Base
         if (YII_ENV === 'test' || mb_strpos($toEmail, '@example.org') !== false) {
             return EMailLog::STATUS_SKIPPED_OTHER;
         }
-        if (EMailBlacklist::isBlacklisted($toEmail)) {
-            return EMailLog::STATUS_SKIPPED_BLACKLIST;
+        if (EMailBlocklist::isBlocked($toEmail)) {
+            return EMailLog::STATUS_SKIPPED_BLOCKLIST;
         }
 
         $message->setTo($toEmail);