+++ migrations/m210116_080438_rename_email_blocklist.php
@@ -1,0 +1,25 @@
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Class m210116_080438_rename_email_blocklist
+ */
+class m210116_080438_rename_email_blocklist extends Migration
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function safeUp()
+    {
+        $this->renameTable('emailBlacklist', 'emailBlocklist');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function safeDown()
+    {
+        $this->renameTable('emailBlocklist', 'emailBlacklist');
+    }
+}